{"schema":"decodex/commit/1","summary":"Extract settings panel sections","authority":"manual"}
